+++ src/include/run_forward.h
@@ -32,7 +32,6 @@ bool runCode(
     ForwardResult* fwdOut                = nullptr,
     const RestartState* warmStart        = nullptr,
     int aWeighting                       = 0,
-    Real ncrithyst                       = 0.2,
     bool verbose                         = false,
     double f_min                         = 200.0,
     double f_max                         = 20000.0,